+++ upcloud/service/managed_database_test.go
@@ -264,49 +264,6 @@ func TestService_GetManagedDatabaseLogs(t *testing.T) {
 	})
 }
 
-func TestService_GetManagedDatabaseConnections(t *testing.T) {
-	record(t, "getmanageddatabaseconnections", func(ctx context.Context, t *testing.T, rec *recorder.Recorder, svc *Service) {
-		createReq := getTestCreateRequest("getmanageddatabaseconnections", upcloud.ManagedDatabaseServiceTypePostgreSQL)
-		createReq.Type = upcloud.ManagedDatabaseServiceTypePostgreSQL
-		createReq.Properties.SetPublicAccess(true).SetIPFilter(upcloud.ManagedDatabaseAllIPv4)
-		serviceDetails, err := svc.CreateManagedDatabase(ctx, createReq)
-		if !assert.NoError(t, err) {
-			return
-		}
-		defer func() {
-			t.Logf("deleting %s", serviceDetails.UUID)
-			err := svc.DeleteManagedDatabase(ctx, &request.DeleteManagedDatabaseRequest{UUID: serviceDetails.UUID})
-			assert.NoError(t, err)
-		}()
-		require.NoError(t, waitForManagedDatabaseRunningState(ctx, rec, svc, serviceDetails.UUID))
-		conns, err := svc.GetManagedDatabaseConnections(ctx, &request.GetManagedDatabaseConnectionsRequest{
-			UUID:   serviceDetails.UUID,
-			Limit:  1000,
-			Offset: 0,
-		})
-		if !assert.NoError(t, err) {
-			return
-		}
-		assert.Len(t, conns, 0)
-
-		err = svc.CancelManagedDatabaseConnection(ctx, &request.CancelManagedDatabaseConnection{
-			UUID:      serviceDetails.UUID,
-			Pid:       0,
-			Terminate: true,
-		})
-		assert.Error(t, err)
-		assert.True(t, strings.HasPrefix(err.(*upcloud.Problem).Title, "Must provide a connection"))
-
-		err = svc.CancelManagedDatabaseConnection(ctx, &request.CancelManagedDatabaseConnection{
-			UUID:      serviceDetails.UUID,
-			Pid:       0,
-			Terminate: false,
-		})
-		assert.Error(t, err)
-		assert.True(t, strings.HasPrefix(err.(*upcloud.Problem).Title, "Must provide a connection"))
-	})
-}
-
 func TestService_GetManagedDatabaseSessions(t *testing.T) {
 	record(t, "getmanageddatabasesessions", func(ctx context.Context, t *testing.T, rec *recorder.Recorder, svc *Service) {
 		createReq := getTestCreateRequest("getmanageddatabasesessions", upcloud.ManagedDatabaseServiceTypePostgreSQL)