+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Removed
+- **Breaking**, Managed Database: connection related methods in favor of session
+
 ## [6.12.0]
 
 ### Added